+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+- Added
+  - Command and menu item for copying URN of an object to clipboard
+
 ## [1.4.0] - 2019-10-22
 
 - Added